fixed some bugs, wrong secondary loop index

+++ pasteRhythm.lua
@@ -8,7 +8,7 @@ loadDependency("preferences")
 loadDependency("midiEditor")
 
 
-local function getExistingNote(existingNotes, startingNotePosition)
+local function getExistingNoteIndex(existingNotes, startingNotePosition)
 
 	for i = 1, #existingNotes do
 
@@ -16,7 +16,7 @@ local function getExistingNote(existingNotes, startingNotePosition)
 		local existingNoteStartingPosition = existingNotes[i][1]
 
 	  if existingNoteStartingPosition == startingNotePosition then
-	  	return existingNote
+	  	return i
 	  end
 	end
 
@@ -34,11 +34,11 @@ local function getExistingNotes(selectedTake)
 
 		local _, noteIsSelected, noteIsMuted, noteStartPositionPPQ, noteEndPositionPPQ, noteChannel, notePitch, noteVelocity = reaper.MIDI_GetNote(selectedTake, noteIndex)
 
-			local existingNote = getExistingNote(existingNotes, noteStartPositionPPQ)
+			local existingNoteIndex = getExistingNoteIndex(existingNotes, noteStartPositionPPQ)
 
-			if existingNote == nil then
+			if existingNoteIndex == nil then
 
-				existingNote = {}
+				local existingNote = {}
 				table.insert(existingNote, noteStartPositionPPQ)
 
 				local existingNoteChannels = {}
@@ -53,13 +53,18 @@ local function getExistingNotes(selectedTake)
 				table.insert(existingNote, existingNoteChannels)
 				table.insert(existingNote, existingNoteVelocities)
 				table.insert(existingNote, existingNotePitches)
+
+				table.insert(existingNotes, existingNote)
 			else
+
+				local existingNote = existingNotes[existingNoteIndex]
+
 				table.insert(existingNote[2], noteChannel)
 				table.insert(existingNote[3], noteVelocity)
 				table.insert(existingNote[4], notePitch)
-			end
 
-			table.insert(existingNotes, existingNote)
+				table.insert(existingNotes[existingNoteIndex], existingNote)
+			end
 	end
 
 	return existingNotes
@@ -122,14 +127,19 @@ local function deleteAllNotes(selectedTake)
 end
 
 
-local function pasteRhythm(mediaItem)
+local function pasteRhythm(rhythmNotes, mediaItem)
 
 	local selectedTake = reaper.GetActiveTake(mediaItem)
+	
+	local existingNotes = getExistingNotes(selectedTake)  
 
-	local rhythmNotes = getRhythmNotesFromPreferences()
-	local existingNotes = getExistingNotes(selectedTake)
+	if #existingNotes == 0 then
+		return
+	end
 
 	deleteAllNotes(selectedTake)
+
+	local previousNoteVelocity
 
 	for i = 1, #rhythmNotes do
 
@@ -138,27 +148,51 @@ local function pasteRhythm(mediaItem)
 		local rhythmNoteStartingPosition = rhythmNote[1][1]
 		local rhythmNoteEndingPosition = rhythmNote[1][2]
 
-		local rhythmNoteChannels = rhythmNote[2]
+		--local rhythmNoteChannels = rhythmNote[2]
 
 		local notePitches = getNearestSetOfNotePitches(existingNotes, rhythmNoteStartingPosition)
 		local noteChannels = getNearestSetOfNoteChannels(existingNotes, rhythmNoteStartingPosition)
 
 		local rhythmNoteVelocities = rhythmNote[3]
 
-		for i = 1, #notePitches do
-			insertMidiNote(selectedTake, rhythmNoteStartingPosition, rhythmNoteEndingPosition, noteChannels[i], notePitches[i], rhythmNoteVelocities[i])
+		for j = 1, #notePitches do
+
+			local velocity = rhythmNoteVelocities[j]
+
+			if velocity == nil then
+				velocity = previousNoteVelocity
+			else
+				previousNoteVelocity = velocity
+			end
+
+			insertMidiNote(selectedTake, rhythmNoteStartingPosition, rhythmNoteEndingPosition, noteChannels[j], notePitches[j], velocity)
 		end
 	end
 end
 
 --
 
+
+reaper.defer(emptyFunctionToPreventAutomaticCreationOfUndoPoint)
+
 local numberOfSelectedItems = getNumberOfSelectedItems()
+
+if numberOfSelectedItems == 0 then
+	return
+end
+
+local rhythmNotes = getRhythmNotesFromPreferences()
+
+if rhythmNotes == nil then
+	return
+end
 
 for i = 0, numberOfSelectedItems-1 do
 
 	local activeProjectIndex = 0
-	local selectedMediaITem = reaper.GetSelectedMediaItem(activeProjectIndex, i)
-	pasteRhythm(selectedMediaITem)
-end
+	local selectedMediaItem = reaper.GetSelectedMediaItem(activeProjectIndex, i)
 
+	startUndoBlock()
+		pasteRhythm(rhythmNotes, selectedMediaItem)
+	endUndoBlock("paste rhythm")
+end
